+++ src/main.tsx
@@ -2,11 +2,11 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { usePDF } from '@react-pdf/renderer';
-import PDF from './pdf';
+import Fancy1 from './variants/fancy-1/pdf';
 import './main.css';
 
 function App() {
-  const [instance] = usePDF({ document: PDF });
+  const [instance] = usePDF({ document: Fancy1 });
 
   if (instance.loading) return <div>Loading...</div>;
   if (instance.error || !instance.url) {
