--- src/projects.ts
@@ -1,62 +1,0 @@
-export default [
-  {
-    id: 'kyleaupton/boot-builder',
-    icons: ['vue', 'electron', 'typescript'],
-    title: 'Boot Builder',
-    description: 'A desktop application to facilitate the creation of bootable USB drives. Built with Vue.js and Electron. Offers a simple, intuitive interface for creating macOS, Windows, and Linux bootable USB installers.',
-  },
-  {
-    id: 'kyleaupton/leetcode-reminder',
-    icons: ['typescript'],
-    title: 'LeetScold',
-    description: 'A Discord bot that reminds users to do their daily LeetCode problem. Uses OpenAI\'s Chat GPT API to generate a daily reminder message, and sends that message to a Discord channel at a specified time each day.',
-  },
-  {
-    id: 'kyleaupton/node-rsync',
-    icons: ['typescript'],
-    title: 'Node Rsync',
-    description: 'A simple rsync wrapper API for server-side JavaScript runtimes. Including a fully typed API with progress and ETA reporting. Easily copy files between servers without the need to shell out to the command line.',
-  },
-  {
-    id: 'kyleaupton/glob-copy',
-    icons: ['typescript'],
-    title: 'Glob Copy',
-    description: 'A simple glob copy utility for Node.js. Copies files matching a glob pattern to a destination directory using Node.js streams. Progress and ETA reporting included.',
-  },
-  {
-    id: 'kyleaupton/portfolio',
-    icons: ['vue', 'typescript'],
-    title: 'Portfolio',
-    description: 'Portfolio website to showcase my projects. Queries GitHub\'s API to dynamically generate project cards. Features a mobile-friendly design, and is built with Vue.js and TypeScript.',
-  },
-  {
-    id: 'kyleaupton/resume',
-    icons: ['react', 'typescript'],
-    title: 'Resume',
-    description: 'My personal resume written in React and programmatically rendered to PDF. This project has a CI/CD pipeline that automatically deploys the latest version of my resume to Amazon S3 whenever a commit is pushed.',
-  },
-  {
-    id: 'kyleaupton/TransactionTracker',
-    icons: ['javascript'],
-    title: 'Transaction Tracker',
-    description: 'A discord bot for tracking Ethereum transactions. Uses various APIs, such as Etherscan and Infura, to alert users when new transactions occur.',
-  },
-  {
-    id: 'kyleaupton/schedule-widget',
-    icons: ['javascript'],
-    title: 'Schedule Widget',
-    description: 'A IOS widget for displaying work schedules. Written in JavaScript using Scriptable.',
-  },
-  {
-    id: 'kyleaupton/win-iso',
-    icons: ['typescript'],
-    title: 'Win ISO',
-    description: 'An API and CLI tool for downloading Windows installation ISOs. Built with TypeScript and Node.js.',
-  },
-  {
-    id: 'kyleaupton/zerotier-utility',
-    icons: ['vue', 'electron', 'javascript'],
-    title: 'ZeroTier Utility',
-    description: 'A desktop application for managing ZeroTier VPN networks. Built with Vue.js and Electron.',
-  },
-];
