--- src/skills.ts
@@ -1,20 +1,0 @@
-export default [
-  'JavaScript',
-  'TypeScript',
-  'HTML',
-  'CSS',
-  'Node.js',
-  'Vue',
-  'Electron',
-  'Fastify',
-  'Express',
-  'Python',
-  'Golang',
-  'Docker',
-  'SQL',
-  'git',
-  'DevOps',
-  'AWS',
-  'Linux',
-  'IP/TCP Networking',
-];
