+++ src/variants/fancy-1/components/Divider.tsx
