+++ src/variants/fancy-1/components/Education.tsx
