+++ src/variants/fancy-1/components/Icons.tsx
