+++ src/variants/fancy-1/components/LineGraph.tsx
