+++ src/variants/fancy-1/components/Link.tsx
