+++ src/variants/fancy-1/components/Project.tsx
