+++ src/variants/fancy-1/components/Projects.tsx
