+++ src/variants/fancy-1/components/Skill.tsx
