+++ src/variants/fancy-1/components/WorkExperience.tsx
