+++ src/variants/fancy-1/pdf.tsx
