+++ src/variants/fancy-1/styles.ts
